Refactor Osm Tag filters
