+++ src/main/java/de/komoot/photon/elasticsearch/PhotonQueryBuilder.java
@@ -4,7 +4,6 @@ package de.komoot.photon.elasticsearch;
 import com.vividsolutions.jts.geom.Envelope;
 import com.vividsolutions.jts.geom.Point;
 import de.komoot.photon.searcher.TagFilter;
-import de.komoot.photon.searcher.TagFilterKind;
 import org.elasticsearch.common.lucene.search.function.CombineFunction;
 import org.elasticsearch.common.lucene.search.function.FiltersFunctionScoreQuery.ScoreMode;
 import org.elasticsearch.common.unit.Fuzziness;
@@ -39,9 +38,7 @@ public class PhotonQueryBuilder {
 
     private State state;
 
-    private BoolQueryBuilder orQueryBuilderForIncludeTagFiltering = null;
-
-    private BoolQueryBuilder andQueryBuilderForExcludeTagFiltering = null;
+    private OsmTagFilter osmTagFilter;
 
     private GeoBoundingBoxQueryBuilder bboxQueryBuilder;
 
@@ -129,6 +126,8 @@ public class PhotonQueryBuilder {
                 .should(QueryBuilders.matchQuery("housenumber", query).analyzer("standard"))
                 .should(QueryBuilders.existsQuery(String.format("name.%s.raw", language)));
 
+        osmTagFilter = new OsmTagFilter();
+        
         state = State.PLAIN;
     }
 
@@ -177,36 +176,8 @@ public class PhotonQueryBuilder {
     }
 
     public PhotonQueryBuilder withOsmTagFilters(List<TagFilter> filters) {
-        for (TagFilter filter : filters) {
-            addOsmTagFilter(filter);
-        }
-        return this;
-    }
-
-    public PhotonQueryBuilder addOsmTagFilter(TagFilter filter) {
         state = State.FILTERED;
-
-        if (filter.getKind() == TagFilterKind.EXCLUDE_VALUE) {
-            appendIncludeTermQuery(QueryBuilders.boolQuery()
-                    .must(QueryBuilders.termQuery("osm_key", filter.getKey()))
-                    .mustNot(QueryBuilders.termQuery("osm_value", filter.getValue())));
-        } else {
-            QueryBuilder builder;
-            if (filter.isKeyOnly()) {
-                builder = QueryBuilders.termQuery("osm_key", filter.getKey());
-            } else if (filter.isValueOnly()) {
-                builder = QueryBuilders.termQuery("osm_value", filter.getValue());
-            } else {
-                builder = QueryBuilders.boolQuery()
-                        .must(QueryBuilders.termQuery("osm_key", filter.getKey()))
-                        .must(QueryBuilders.termQuery("osm_value", filter.getValue()));
-            }
-            if (filter.getKind() == TagFilterKind.INCLUDE) {
-                appendIncludeTermQuery(builder);
-            } else {
-                appendExcludeTermQuery(builder);
-            }
-        }
+        osmTagFilter.withOsmTagFilters(filters);
         return this;
     }
 
@@ -219,7 +190,6 @@ public class PhotonQueryBuilder {
     }
 
 
-
     /**
      * When this method is called, all filters are placed inside their containers and the top level filter
      * builder is built. Subsequent invocations of this method have no additional effect. Note that after this method
@@ -230,12 +200,8 @@ public class PhotonQueryBuilder {
 
         finalQueryBuilder = QueryBuilders.boolQuery().must(finalQueryWithoutTagFilterBuilder).filter(queryBuilderForTopLevelFilter);
 
-        if (state.equals(State.FILTERED)) {
-            BoolQueryBuilder tagFilters = QueryBuilders.boolQuery();
-            if (orQueryBuilderForIncludeTagFiltering != null)
-                tagFilters.must(orQueryBuilderForIncludeTagFiltering);
-            if (andQueryBuilderForExcludeTagFiltering != null)
-                tagFilters.mustNot(andQueryBuilderForExcludeTagFiltering);
+        BoolQueryBuilder tagFilters = osmTagFilter.getTagFiltersQuery();
+        if (state.equals(State.FILTERED) && tagFilters != null) {
             finalQueryBuilder.filter(tagFilters);
         }
         
@@ -249,25 +215,6 @@ public class PhotonQueryBuilder {
 
         return finalQueryBuilder;
     }
-
-
-    private void appendIncludeTermQuery(QueryBuilder termQuery) {
-
-        if (orQueryBuilderForIncludeTagFiltering == null)
-            orQueryBuilderForIncludeTagFiltering = QueryBuilders.boolQuery();
-
-        orQueryBuilderForIncludeTagFiltering.should(termQuery);
-    }
-
-
-    private void appendExcludeTermQuery(QueryBuilder termQuery) {
-
-        if (andQueryBuilderForExcludeTagFiltering == null)
-            andQueryBuilderForExcludeTagFiltering = QueryBuilders.boolQuery();
-
-        andQueryBuilderForExcludeTagFiltering.should(termQuery);
-    }
-
 
     private enum State {
         PLAIN, FILTERED, QUERY_ALREADY_BUILT, FINISHED,
